Enable custom math verification.
For ilogb(0.0f), we want to allow two possible values.  This does
not match the current pattern where the CoreMathVerifier returns
the expected outputs given specified inputs.

This introduces a "custom" test mode where the Test* method calls the
CoreMathVerifier, passing it both the actual inputs and outputs, and
asking these to be verified.

Change-Id: I2e4b0201c93b206e3eb48a6832519f8b827f969f