[MIPS] Add the got table right below the text section
Make sure the got table is close to the text section, so the function
can reach it easily. This is important for MIPS64 for which there is an
assumption that got table can be reached if 32-bit value is added to $t9
(the start of the function for PIC). The MIPS64 code looks like this:

  // lui $v0, %hi(%neg(%gp_rel(fname)))
  // daddu $v1, $v0, $t9
  // daddiu $globalbasereg, $v1, %lo(%neg(%gp_rel(fname)))

This patch prevents possible issue in which a function could not reach
the got table.

Change-Id: I024f1f9483ffcd312bb6fcbc61145a171e7096e9