Initial python test suite for debugging a RenderScript app using lldb.
Includes testing for newly introduced lldb commands, which are subcommands of "language renderscript", including tests for setting a breakpoint on a kernel or a filename:line, on Java apps containing the wait-for-attach flag as well as an infinite loop (hence testing attaching to a running target). Also included are the sources to 6 binaries for current and future use: Java and NDK versions with and without debug information of a test with the wait-for-attach flag and with an infinite loop. Currently only 32bit ARM targets are officially supported (i.e. have verified 100% pass rate with appropriate patches).

Change-Id: Ib3d29fdc02aa113028a49b4e9e86a50f0abd7fac
Signed-off-by: verena beckham <verena@codeplay.com>