Don't use anonymous literal structures for RS object types.
Bug: 22926131
Bug: 23445219

Using an anonymous structure type prevents us from being able to use
LLVM to determine the actual typename. This is problematic, because
there are some cases where bcc needs to be able to detect/act on RS
object types. Giving each of these structures a legitimate name has no
impact on the generated code (since we are already using typedefs).
This change is also safe for targeting prior Android releases with our
toolchain.

Change-Id: Iaff928f5821af66cfc9b3aea2ff5549d0c8a9408